delete initStructEditor and move code to StructEditor. change import in cel-table
